feat: separates workflow publishing and sets all workflows to be published
